+++ R/tree_generation.R
@@ -35,7 +35,6 @@ NULL
 #' data('Lobo')
 #' RandomTree(Lobo.phy)
 #'
-#' @importFrom ape root
 #' @export
 RandomTree <- function (tips, root = FALSE) {
   tips <- TipLabels(tips)
@@ -43,6 +42,17 @@ RandomTree <- function (tips, root = FALSE) {
   edge <- do.call(cbind,
                   RenumberEdges(.RandomParent(nTips),
                                 seq_len(nTips + nTips - 2L)))
+  if (!is.logical(root) && root != 1L) {
+    if (isTRUE(root)) root <- 1L
+    if (is.character(root)) root <- which(tips == root)
+    if (length(root) == 0L) stop ("No match found for `root`")
+    if (!is.integer(root)) root <- as.integer(root)
+    if (length(root) > 1L) {
+      root <- root[1]
+      warning("More than one entry in `root`; using ", root)
+    }
+    edge <- root_on_node(edge, root)
+  }
   tree <- structure(list(edge = edge,
                          Nnode = nTips - 1L,
                          tip.label = tips,
@@ -50,10 +60,6 @@ RandomTree <- function (tips, root = FALSE) {
                     class = 'phylo')
   if (isFALSE(root)) {
     tree <- UnrootTree(tree)
-  } else if(isTRUE(root) || root == 1L) {
-    # Tree already rooted on '1'
-  } else {
-    tree <- root(tree, root, resolve.root = TRUE)
   }
 
   # Return: