+++ src/RcppExports.cpp
@@ -55,25 +55,37 @@ BEGIN_RCPP
 END_RCPP
 }
 // preorder_edges_and_nodes
-IntegerMatrix preorder_edges_and_nodes(IntegerVector parent, IntegerVector child);
+IntegerMatrix preorder_edges_and_nodes(const IntegerVector parent, const IntegerVector child);
 RcppExport SEXP _TreeTools_preorder_edges_and_nodes(SEXP parentSEXP, SEXP childSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< IntegerVector >::type parent(parentSEXP);
-    Rcpp::traits::input_parameter< IntegerVector >::type child(childSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector >::type parent(parentSEXP);
+    Rcpp::traits::input_parameter< const IntegerVector >::type child(childSEXP);
     rcpp_result_gen = Rcpp::wrap(preorder_edges_and_nodes(parent, child));
     return rcpp_result_gen;
 END_RCPP
 }
 // postorder_edges
-IntegerMatrix postorder_edges(IntegerMatrix edge);
+IntegerMatrix postorder_edges(const IntegerMatrix edge);
 RcppExport SEXP _TreeTools_postorder_edges(SEXP edgeSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< IntegerMatrix >::type edge(edgeSEXP);
+    Rcpp::traits::input_parameter< const IntegerMatrix >::type edge(edgeSEXP);
     rcpp_result_gen = Rcpp::wrap(postorder_edges(edge));
+    return rcpp_result_gen;
+END_RCPP
+}
+// root_on_node
+IntegerMatrix root_on_node(const IntegerMatrix edge, const int outgroup);
+RcppExport SEXP _TreeTools_root_on_node(SEXP edgeSEXP, SEXP outgroupSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const IntegerMatrix >::type edge(edgeSEXP);
+    Rcpp::traits::input_parameter< const int >::type outgroup(outgroupSEXP);
+    rcpp_result_gen = Rcpp::wrap(root_on_node(edge, outgroup));
     return rcpp_result_gen;
 END_RCPP
 }