+++ NEWS.md
@@ -1,7 +1,12 @@
 # TreeTools 1.2.0.9010 (development)
 
- - Export `preorder_edges_and_nodes()` C++ function.
- 
+ - Export C++ functions `preorder_edges_and_nodes()` and `postorder_edges()`.
+ - New C++ function `root_on_node()` roots binary trees.
+   # To do: 
+     [ ] document
+     [ ] integrate into R funcs
+     [X] write test cases
+
 # TreeTools 1.2.0.9005 (development)
 
  - Import RdMacros package 'Rdpack'.